+++ tests/testthat.R
@@ -1,3 +1,5 @@
+#' MassIVE ID: MSV000080547
+
 library(testthat)
 library(MsBackendMassIVE)
 
@@ -5,7 +7,8 @@ test_check("MsBackendMassIVE")
 
 ## Run tests with the unit test suite defined in the Spectra package to ensure
 ## compliance with the definitions of the MsBackend interface/class.
-be <- ...
+be <- backendInitialize(MsBackendMassIVE(), "MSV000080547",
+                        filePattern = "1.mzML$")
 
 library(Spectra)
 test_suite <- system.file("test_backends", "test_MsBackend",