+++ NEWS.md
@@ -1,1 +1,5 @@
-# MsBackendMassIVE 0.0
+# MsBackendMassIVE 0.1
+
+- Add MassIVE utilities
+- Add functionality to download and cache MS data files from MassIVE
+- Add unit tests